% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FusionForest.R
\name{FusionShrinkageForest}
\alias{FusionShrinkageForest}
\title{Fusion Forests}
\usage{
FusionShrinkageForest(
  y,
  status = NULL,
  X_train_control,
  X_train_treat,
  treatment_indicator_train,
  source_indicator_train,
  X_test_control = NULL,
  X_test_treat = NULL,
  X_test_deconf = NULL,
  treatment_indicator_test = NULL,
  source_indicator_test = NULL,
  outcome_type = "continuous",
  timescale = "time",
  number_of_trees_control = 200,
  number_of_trees_treat = 200,
  number_of_trees_deconf = 200,
  prior_type_control = "horseshoe",
  prior_type_treat = "horseshoe",
  prior_type_deconf = "horseshoe",
  local_hp_control,
  local_hp_treat,
  local_hp_deconf,
  global_hp_control = NULL,
  global_hp_treat = NULL,
  global_hp_deconf = NULL,
  power = 2,
  base = 0.95,
  p_grow = 0.4,
  p_prune = 0.4,
  nu = 3,
  q = 0.9,
  sigma = NULL,
  rho = 1,
  rho_known = TRUE,
  N_post = 5000,
  N_burn = 5000,
  delayed_proposal = 5,
  store_posterior_sample = FALSE,
  store_posterior_sample_deconf = FALSE,
  seed = NULL,
  verbose = TRUE
)
}
\description{
TBD
}
