% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CausalShrinkageForest.R
\name{CausalShrinkageForest}
\alias{CausalShrinkageForest}
\title{General Causal Shrinkage Forests}
\usage{
CausalShrinkageForest(
  y,
  status = NULL,
  X_control_train,
  X_treat_train,
  treatment_indicator_train,
  X_control_test = NULL,
  X_treat_test = NULL,
  treatment_indicator_test = NULL,
  number_of_trees_control = 200,
  number_of_trees_treat = 200,
  N_post = 5000,
  N_burn = 5000,
  delayed_proposal = 5,
  power = 2,
  base = 0.95,
  p_grow = 0.4,
  p_prune = 0.4,
  nu = 3,
  q = 0.9,
  alpha_local_control,
  alpha_local_treat,
  alpha_global_control = NULL,
  alpha_global_treat = NULL,
  tau_control = NULL,
  tau_treat = NULL,
  forest_wide_shrinkage_control = FALSE,
  forest_wide_shrinkage_treat = FALSE,
  sigma = NULL,
  y_mean = NULL,
  store_posterior_sample_control = FALSE,
  store_posterior_sample_treat = FALSE,
  seed = NULL,
  scale = "time",
  verbose = TRUE
)
}
\description{
DESCRIPTION TBD.
}
