% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShrinkageTrees-package.R
\docType{package}
\name{ShrinkageTrees-package}
\alias{ShrinkageTrees-package}
\title{ShrinkageTrees: Bayesian Tree Ensembles for Survival Analysis and Causal Inference}
\description{
Bayesian regression tree ensembles for survival analysis and causal inference. Implements BART, DART, Bayesian Causal Forests (BCF), and Horseshoe Forest models. Supports right-censored and interval-censored survival outcomes via accelerated failure time (AFT) formulations. Designed for high-dimensional prediction and heterogeneous treatment effect estimation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tijn-jacobs/ShrinkageTrees}
  \item Report bugs at \url{https://github.com/tijn-jacobs/ShrinkageTrees/issues}
}

}
\author{
\strong{Maintainer}: Tijn Jacobs \email{t.jacobs@vu.nl} (\href{https://orcid.org/0009-0003-6188-9296}{ORCID})

}
\keyword{internal}
