% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurvivalWrappers.R
\name{SurvivalBCF}
\alias{SurvivalBCF}
\title{SurvivalBCF (Bayesian Causal Forest; Hahn et al. 2020)}
\usage{
SurvivalBCF(
  time,
  status,
  X_train,
  treatment,
  timescale = "time",
  propensity = NULL,
  N_post = 1000,
  N_burn = 1000,
  verbose = TRUE
)
}
\arguments{
\item{status}{Event indicator (1 = event, 0 = censored). Required for survival.}

\item{X_train}{Design matrix for training.}

\item{N_post}{Number of posterior samples to draw.}

\item{N_burn}{Number of burn-in iterations.}

\item{verbose}{Print sampling progress.}

\item{W}{Treatment indicator (0/1) for training.}

\item{number_of_trees_mu}{Trees for control forest.}

\item{number_of_trees_tau}{Trees for treatment effect forest.}

\item{k_mu, k_tau}{Leaf prior scales.}
}
\description{
A simplified wrapper for fitting an AFT version of BCF:
Y = f(x) + a * τ(x) + ε, with separate control/treatment forests.
}
