% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurvivalWrappers.R
\name{SurvivalShrinkageBCF}
\alias{SurvivalShrinkageBCF}
\title{SurvivalShrinkageBCF (BCF + Dirichlet Sparsity)}
\usage{
SurvivalShrinkageBCF(
  time,
  status,
  X_train,
  treatment,
  timescale = "time",
  propensity = NULL,
  a_dir = 0.5,
  b_dir = 1,
  N_post = 1000,
  N_burn = 1000,
  verbose = TRUE
)
}
\arguments{
\item{status}{Event indicator (1 = event, 0 = censored). Required for survival.}

\item{X_train}{Design matrix for training.}

\item{N_post}{Number of posterior samples to draw.}

\item{N_burn}{Number of burn-in iterations.}

\item{verbose}{Print sampling progress.}

\item{prior_mu, prior_tau}{Prior types for control & treatment forests.}
}
\description{
This model generalizes BCF by including:
\itemize{
\item Dirichlet splitting priors for sparsity (optional).
}
}
